Fix(cli): improve the help messages for the cli

+++ src/main.rs
@@ -12,10 +12,12 @@ use log::Level;
 use loggerv::{Logger, Output};
 use structopt::StructOpt;
 
+/// Run sementic-release steps in the context of a cargo based Rust project.
 #[derive(StructOpt)]
 struct Opt {
+    /// Increases the logging level (use multiple times for more detail).
     #[structopt(short, long, parse(from_occurrences))]
-    verbose: u8,
+    verbose: u64,
 }
 
 fn main() -> Result<(), Error> {
@@ -26,7 +28,7 @@ fn main() -> Result<(), Error> {
         .output(&Level::Trace, Output::Stderr)
         .output(&Level::Debug, Output::Stderr)
         .output(&Level::Info, Output::Stderr)
-        .verbosity(opt.verbose as u64)
+        .verbosity(opt.verbose)
         .init()?;
 
     println!("Hello, world!");
